Lv 3-2) Calculator을 제네릭 클래스로. 하지만 double로 형변환을 일일이 해야해서 이게 맞는지 모르겠음..

+++ src/main/java/calculator/DivideOperator.java
@@ -3,8 +3,8 @@ package calculator;
 public class DivideOperator implements Operator{
     double result;
     @Override
-    public double operate(int num1, int num2){
-        result =(double)num1 / num2;
+    public double operate(double num1, double num2){
+        result =num1 / num2;
         return result;
     }
 }
